# doc-specs/rest-1/rest-plugin-openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Aquila-Plugin-Portal",
    "version" : "1.0.0",
    "contact" : {
      "name" : "Kilian Krampf",
      "email" : "kilian@mail.menkalian.de",
      "url" : "https://plugins.aquila.menkalian.de/contact"
    },
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "description" : "API for the Aquila Plugin-Portal Backend.\nAuthentication works via OAuth (with additional registration at the server).\nTHIS IS ONLY THE MINIMUM REQUIRED API! Individual implementations may expand upon this.",
    "termsOfService" : "https://plugins.aquila.menkalian.de/terms"
  },
  "paths" : {
...
    "/user" : {
      "summary" : "Operations regarding the user",
      "get" : {
        "summary" : "",
        "description" : "",
        "parameters" : [
          {
            "name" : "id",
            "in" : "query",
            "required" : true,
            "schema" : {
              "type" : "integer"
            },
            "description" : "ID of the updated plugin"
          }
        ],
        "operationId" : "getUserById",
        "responses" : {
          "default" : {
            "description" : "Default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserData"
                }
              }
            }
          }
        },
        "tags" : [
          "authentication"
        ],
        "security" : []
      },
      "post" : {
        "summary" : "",
        "description" : "",
        "parameters" : [],
        "operationId" : "updateCurrentUser",
        "responses" : {
          "default" : {
            "description" : "Default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserData"
                }
              }
            }
          }
        },
        "tags" : [
          "authentication"
        ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserUploadData"
              }
            }
          },
          "description" : "New data for the current user (id is not changeable and will be ignored)"
        }
      },
      "put" : {
        "summary" : "",
        "description" : "",
        "parameters" : [],
        "operationId" : "createNewUser",
        "responses" : {
          "default" : {
            "description" : "Default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserData"
                }
              }
            }
          }
        },
        "tags" : [
          "authentication"
        ],
        "requestBody" : {
          "required" : false,
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserUploadData"
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Deletes current user",
        "description" : "Deletes the current user and archives all associated plugins.\nThe associated plugins will be transferred to an annonymous account and no one will be able to change their data.",
        "parameters" : [],
        "operationId" : "deleteCurrentUser",
        "responses" : {
          "default" : {
            "description" : "Default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "boolean"
                }
              }
            }
          }
        },
        "tags" : [
          "authentication"
        ]
      }
    },
...
  },
...
  "security" : [
    {
      "googleOAuth" : []
    }
  ],
  "servers" : [
    {
      "url" : "plugin.aquila.menkalian.de",
      "description" : "production server with actual data",
      "variables" : {}
    },
    {
      "url" : "dev.plugin.aquila.menkalian.de",
      "description" : "test server (for testing your integration with dummy data)"
    },
    {
      "url" : "http://localhost:8080",
      "description" : "local server for testing"
    }
  ],
  "components" : {
...
    "schemas" : {
      "UserData" : {
        "type" : "object",
        "required" : [
          "userId",
          "name"
        ],
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer"
          },
          "avatarUrl" : {
            "type" : "string"
          }
        }
      },
      "UserUploadData" : {
        "type" : "object",
        "required" : [
          "userId",
          "name"
        ],
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer"
          },
          "avatarData" : {
            "type" : "array",
            "description" : "Image data of the new avatar to use. If not present, the avatar should be deleted. If set to an empty array (len = 0), the current value remains unchanged.",
            "items" : {
              "type" : "integer"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "googleOAuth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl" : "https://www.googleapis.com/oauth2/v4/token",
            "refreshUrl" : "https://www.googleapis.com/oauth2/v4/token",
            "scopes" : {}
          }
        }
      }
    }
  },
...
}